requirements.txt
fastapi
uvicorn
google-cloud-storage
matplotlib
pandas
python-multipart
starlette
